+++ src/buffer_attachment.rs
@@ -1,36 +1,36 @@
 #[derive(Debug, Clone)]
 pub struct BufferAttachment<T> {
     buffer: Vec<T>,
-    width: usize,
-    height: usize,
+    width: i32,
+    height: i32,
 }
 
 impl<T> BufferAttachment<T>
 where
     T: Clone,
 {
-    pub fn new(width: usize, height: usize, value: T) -> Self {
+    pub fn new(width: i32, height: i32, value: T) -> Self {
         Self {
-            buffer: vec![value; width * height],
+            buffer: vec![value; (width * height) as usize],
             width: width,
             height: height,
         }
     }
 
-    pub fn width(&self) -> usize {
+    pub fn width(&self) -> i32 {
         self.width
     }
 
-    pub fn height(&self) -> usize {
+    pub fn height(&self) -> i32 {
         self.height
     }
 
-    pub fn set(&mut self, x: usize, y: usize, value: T) {
-        self.buffer[y * self.width + x] = value;
+    pub fn set(&mut self, x: i32, y: i32, value: T) {
+        self.buffer[(y * self.width + x) as usize] = value;
     }
 
-    pub fn get(&mut self, x: usize, y: usize) -> &mut T {
-        &mut self.buffer[y * self.width + x]
+    pub fn get(&mut self, x: i32, y: i32) -> &mut T {
+        &mut self.buffer[(y * self.width + x) as usize]
     }
 
     pub fn buffer(&self) -> &Vec<T> {
